feat!: derived-inputs-outputs rule (#25)
Adding a new rule to ensure that the values passed to the callback of a derived() always match the names of the input stores with a $ prefixed to them.

BREAKING CHANGE:

`stores-no-async` rule will no longer auto-fix and remove the leading `async` keyword, instead it will make a suggestion to that effect. The previous behavior wasn't safe.